+++ cheadergen/tests/config/fail/opaque_on_constant.rs
@@ -1,4 +1,4 @@
-#[cheadergen::config(export(opaque))]
+#[cheadergen::config(opaque)]
 pub const FOO: u32 = 1;
 
 fn main() {}